backend/requirements.txt
Flask==2.3.2
gunicorn==21.2.0
flask-cors==3.0.10
SQLAlchemy==2.0.23
alembic==1.13.1
Werkzeug==3.0.1
PyJWT==2.8.0
